find todos in main RepoBranchesRepository, consolidate slack helpers
